backend/requirements.txt
asgiref
certifi
chardet
coreapi
coreschema
Django
django-comment-migrate
django-cors-headers
django-filter
django-ranged-response
django-restql
django-simple-captcha
django-timezone-field
djangorestframework
djangorestframework-simplejwt
drf-yasg
idna
inflection
itypes
Jinja2
MarkupSafe
mysqlclient
packaging
Pillow
PyJWT
pyparsing
pyPEG2
pypinyin
pytz
requests
ruamel.yaml
ruamel.yaml.clib
six
smmap
sqlparse
typing-extensions
tzlocal
ua-parser
uritemplate
urllib3
user-agents
whitenoise
openpyxl
uvicorn
gunicorn
gevent
websockets
channels
channels-redis